amd64 symantic stubs
