+++ lib/CMakeLists.txt
@@ -2,7 +2,8 @@ set(panopticon_SOURCES src/loc.cc src/value.cc src/region.cc src/marshal.cc
 	src/database.cc src/structure.cc src/basic_block.cc src/mnemonic.cc
 	src/procedure.cc src/program.cc src/disassembler.cc src/dflow.cc
 	src/instr.cc src/interpreter.cc src/avr/avr.cc src/avr/util.cc
-	src/code_generator.cc src/amd64/amd64.cc src/amd64/decode.cc src/amd64/generic.cc)
+	src/code_generator.cc src/amd64/traits.cc src/amd64/decode.cc
+	src/amd64/symantic.cc)
 set(panopticon_HEADERS include/panopticon/loc.hh include/panopticon/value.hh
 	include/panopticon/marshal.hh include/panopticon/digraph.hh
 	include/panopticon/region.hh include/panopticon/hash.hh
@@ -15,7 +16,8 @@ set(panopticon_HEADERS include/panopticon/loc.hh include/panopticon/value.hh
 	include/panopticon/avr/util.hh include/panopticon/architecture.hh
 	include/panopticon/ensure.hh include/panopticon/tree.hh
 	include/panopticon/amd64/amd64.hh include/panopticon/amd64/decode.hh
-	include/panopticon/amd64/generic.hh)
+	include/panopticon/amd64/syntax.hh include/panopticon/amd64/symantic.hh
+	include/panopticon/amd64/traits.hh)
 
 set(panopticon_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/include
 	${KyotoCabinet_INCLUDE_DIRS} ${LibArchive_INCLUDE_DIRS}
